deployment: Support multiple passes of connecting data ports
The first pass must be tolerant of connections that only have one
port currently - future ports may complete this connection.

So during the first pass, when loading components only ports that
are compiled in to a component (ie created in the component's
constructor), it is acceptable to have a connection with only one
port and such a connection is ignored for now.

During the second pass, after configuring components (and thus any
ports created during configureHook(), say in response to properties)
at this point all ports should have been created and a connection
with only one port shall have a stream created for it.