deployment: set default connection policy from XML deployment file
If a ConnPolicy element with the special name "Default" exists in the file,
it will overwrite the default for subsequent new connections. There might
be multiple ConnPolicy elements named Default in the file, and each of them
only affect the defaults for connection created later in the file.

A new connection is created whenever the parser sees a named ConnPolicy element or
when a connection with an unknown name is read as a value in the "Ports" part of a
"Component" element.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>