deployment: accept name "this" as an alias for the DeploymentComponent in arguments (fix #40)
Because the DeploymentComponent overrides the loadService operation and the name of the deployer
might not be known, the operations which translate a string into a TaskContext should accept
"this" as an alias for the DeploymentComponent itself. This is in line with the usage of the
"this" keyword in RTT scripting.

Use case:

import("rtt_rosdeployment")
loadService("this", "rosdeployment")

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>