logging: Remove unused RT string from real-time LoggingEvent structure
The "ndc" value is not used in the real-time logging framework as it is
inherently non-real-time safe in log4cpp itself. Instead an empty string
is passed through. To prevent the repetitive use of an empty string,
just don't use this at all and pass the empty string explicitly back
to log4cpp when converting to log4cpp in toLog4cpp().