logging: Change Appender's logging port to non-event port
An event port creates a direction connection between the logging
entity/category, and the appender. This connection involves a mutex
that can create a priority inversion situation (with a lack of priority
inheritance involved). Alternatively, even with priority inheritance
enabled this can cause a low-priority appending task to block a
higher-priority task that has already completed the act of pushing
the logging event into the buffer.

The fix is to make this a non-event port, and put the burden on the
appender (ie the consumer) to wake itself up appropriately to consume
logging events sufficiently quickly to not lose any.